Document properly and fixes various bugs
