+++ src/privacy.rs
@@ -2,6 +2,7 @@ use crate::templates::privacy_template::PrivacyTemplate;
 use actix_web::Error;
 use actix_web::Responder;
 
+// Returns the privacy settings site
 pub async fn privacy() -> Result<impl Responder, Error> {
     Ok(PrivacyTemplate {})
 }
