+++ src/state.rs
@@ -7,6 +7,7 @@ pub enum State {
     None,
 }
 
+/// utility for displaying the state without manual matching
 impl Display for State {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let s = match self {
