+++ templates/dashboard.html
@@ -11,6 +11,11 @@
             hx-swap="innerHTML" hx-vals='{"key": "{{ key }}"}'>
             Change State
         </button>
+
+        <button class="bg-red-600 hover:bg-red-700 px-3 py-1 rounded" hx-post="/remove_state" hx-target="body"
+            hx-swap="innerHTML" hx-vals='{"key": "{{ key }}"}'>
+            Change State
+        </button>
     </li>
     {% endfor %}
 </ul>
